:rose: fix minor bugs in error handlers

+++ src/simmate/calculators/vasp/error_handlers/insufficient_bands.py
@@ -37,7 +37,7 @@ class InsufficientBands(ErrorHandler):
                     nbands_current = int(line.split()[-1])
                     break
         # increase the number of bands by 10%
-        nbands_new = nbands_current * 1.1
+        nbands_new = int(nbands_current * 1.1)
         incar["NBANDS"] = nbands_new
         correction = f"switch NBANDS from {nbands_current} to {nbands_new} (+10%)"
